Fail and notify the user if the docker desktop CLI cannot run
If docker cannot be found on the PATH then the docker desktop CLI
command will not complete. In that case our code is stuck because the
promise is never resolved. Catching the error case where docker cannot
be found on the PATH and resolving the promise will ensure that the user
is prompted to install Docker Desktop.

Signed-off-by: Remy Suen <remy.suen@docker.com>