Register the Scout command regardless of any user setting
If the rollout setting is set to featureFlag and the feature flag is
currently false then the Scout command will fail. However, if the
setting is set to disabled or enabled then the Scout command will work.
As the feature flag is currently set to false, users will see the
command expecting it to work when it will not. Explicitly registering
the command regardless of any feature flag or setting will help resolve
this issue.

Signed-off-by: Remy Suen <remy.suen@docker.com>