Check for Docker Desktop correctly
When the code was refactored to use a generic spawnDockerCommand
function, the removal of 'docker' was missed so we were no longer
running 'docker desktop version' correctly. We also need to use the
onError function to ensure that the code is executed correctly in the
event that the docker binary cannot be found.

Signed-off-by: Remy Suen <remy.suen@docker.com>