Change how Compose features are exposed
Previously, we would present a toast to the user to notify them that
they would be seeing duplicates if Red Hat's YAML extension was
installed. We will now auto-disable the Compose features if the YAML
extension is installed.

If the user does some digging into the settings, they can read the
description and learn more about the situation from the FAQ in our
README.md. They can then proceed to opt-in to our Compose features
manually with the full understanding of the implications of turning it
on. If they follow the steps, they will also be able to disable the
duplicates from Red Hat's extension.

The intent here is to give users full control over their Compose editing
experience in Visual Studio Code. If they want to use Red Hat's YAML
language server, they can. And if they want to use ours, they can.
Whichever one they choose, there will be documentation available to opt
out of the other so that duplicates will not appear.

Signed-off-by: Remy Suen <remy.suen@docker.com>