Include the Dockerfile Language Server in the extension
There are two language servers for Dockerfiles. There is the Dockerfile
Language Server written in TypeScript and the Docker Language Server
written in Go (which also supports Compose files and Bake files). The
Dockerfile Language Server provides features that the Docker Language
Server does not provide so bundling it in the extension will improve the
overall editing experience of Dockerfiles.

If Microsoft's Docker extension or the Container Tools extension is
installed, the Dockerfile Language Server will not be started. This is
because Microsoft's extensions already include the Dockerfile Language
Server. If we also started the Dockerfile Language Server on our side
then users may get duplicated information which we want to avoid.

Signed-off-by: Remy Suen <remy.suen@docker.com>