debugger: update debugger components
Debugger options have been updated to reflect the current version on
master more closely.

The names have been renamed to `Docker: Build` instead of `Dockerfile
Debug`. The reason for this is to make the naming more consistent and
make it so the bake target could be called `Docker: Bake`.

Initial configurations and configuration snippets now more closely align
with the default attributes most projects should expect.

The vscode extension host debugging now uses `test/resources` as the
workspace folder and included a default dockerfile that can be used for
testing debugging.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>