dap: add stopOnEntry attribute to debugger configuration attributes
This adds `stopOnEntry` to the debugger configuration attributes which
will configure whether the debugger stops after the first instruction or
if it continues until the first exception.

This also adds a `contributes.breakpoints` section so vscode will allow
breakpoints to be set.

Some additional activation events have also been added to ensure the
plugin activates even if a dockerfile isn't opened.

The `BUILDX_EXPERIMENTAL` environment variable has been restored because
the pull request planning to remove that hasn't been merged at this
moment so it is still required.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>