Improve launch configuration experience
The dockerfile attribute needs to be a relative path to the working
directory so the description has been corrected. Although it is unlikely
anyone needs to set the working directory for the debug adapter to run
in, there is nothing to lose from exposing this attribute given that the
dockerfile attribute has to be relative to it.

Signed-off-by: Remy Suen <remy.suen@docker.com>