Change the sample launch configuration to be a full launch.json
Someone unfamiliar with the launch.json syntax may copy/paste the code
as-is which will not follow VS Code's schema. Providing a full sample of
a launch.json should help prevent that. For those that already have a
launch.json, they can copy/paste the configuration and merge it into
their existing file.

Signed-off-by: Remy Suen <remy.suen@docker.com>